add AnnotatedCache interface
