+++ src/initialize.ts
@@ -1,7 +1,9 @@
 import { container as defaultContainer } from "./container/container";
 import { keys } from "./container/keys";
 import { getGlobal } from "./container/utils/getGlobal";
-import { AnnotatedRouter, Class, RequestHandler } from "./decorators/types";
+import { Class } from "./decorators/types";
+import { AnnotatedRouter } from "./interfaces/router";
+import { RequestHandler } from "./interfaces/types";
 
 /**
  * Initialize the framework
