+++ src/interfaces/types.ts
@@ -1,0 +1,1 @@
+export type RequestHandler = (req: Request) => Promise<Response>;
